+++ packages/main-ui/src/pages/mod.rs
@@ -5,8 +5,8 @@ mod highlighted_topic_component;
 mod i18n;
 pub mod page;
 mod upcoming_topics_component;
-pub mod politician;
+pub mod politicians;
 
 pub use _routes::*;
 pub use page::*;
-pub use politician::*;
+pub use politicians::*;
